@@ -2,6 +2,7 @@
   "name": "elsie-bray-bot",
   "version": "1.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -27,8 +28,8 @@
     "@types/node": "^24.0.0",
     "@types/react": "^19.1.0",
     "@types/react-dom": "^19.1.0",
-    "eslint": "^8.15.0",
-    "eslint-config-next": "^13.0.4",
+    "eslint": "^9.31.0",
+    "eslint-config-next": "^15.4.1",
     "prettier": "^3.6.2",
     "typescript": "^5.0.4"
   }
