@@ -1,4 +1,4 @@
-const { withSentryConfig } = require('@sentry/nextjs')
+import { withSentryConfig } from '@sentry/nextjs'
 
 const dsn = process.env.SENTRY_DSN || process.env.NEXT_PUBLIC_SENTRY_DSN
 
@@ -37,5 +37,4 @@ const nextConfig = {
   },
 }
 
-// @ts-ignore
-module.exports = withSentryConfig(nextConfig, { dryRun: !dsn, silent: true })
+export default withSentryConfig(nextConfig, { dryRun: !dsn, silent: true })
